+++ packages/panels/resources/views/pages/page.blade.php
@@ -4,7 +4,11 @@
 <div class="wire-resource-page space-y-4 sm:space-y-6">
     @include('wire-panels::pages.partials.header')
 
+    @include('wire-panels::pages.partials.page-widgets', ['pageWidgets' => $headerWidgets])
+
     @include($contentView)
+
+    @include('wire-panels::pages.partials.page-widgets', ['pageWidgets' => $footerWidgets])
 
     @include('wire-panels::pages.partials.action-modals')
 </div>